Reorganized code to better match original
Simple Fission Reactor now somewhat works

Added More Fission Parts
Fixed turbine being too bright
Fixed multiblock frame state not working

+++ src/main/java/com/nred/nuclearcraft/multiblock/fisson/solid/FissionHeatSinkType.java
@@ -67,7 +67,7 @@ public enum FissionHeatSinkType implements StringRepresentable, ITileEnum<SolidF
         return getSerializedName();
     }
 
-    public double getCoolingRate() {
+    public int getCoolingRate() {
         return coolingRate.get();
     }
